+++ src/queries/globalPrefetcher.ts
